LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-pannernode-interface/panner-rolloff-clamping-expected.txt

CONSOLE MESSAGE: line 70: SyntaxError: The string did not match the expected pattern.

Harness Error (FAIL), message = SyntaxError: The string did not match the expected pattern.

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [linear-clamp-low]  
